requirements.txt
Flask
gunicorn
flask-cors
psycopg2-binary
requests
python-dotenv
